% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_cluster_configs}
\alias{get_cluster_configs}
\alias{get_yaml_configuraions_path}
\title{Get cluster configurations}
\usage{
get_cluster_configs(cluster_name, scope, yaml_path)

get_yaml_configuraions_path(cluster_name)
}
\arguments{
\item{cluster_name}{Character. The name of the cluster. Capitolization and spacing
is ignored and does not matter (e.g., "My Cluster" is equivalent to "mycluster")}

\item{scope}{Character. The scope of the configurations of interest. The configuration
YAML may be organized by various headings.}
}
\description{
Find and read the cluster configuration file for a named cluster.
}

