% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_java_classpath}
\alias{get_java_classpath}
\alias{get_java_dependencies_path}
\title{Get java classpath}
\usage{
get_java_classpath(cluster_name, jars)

get_java_dependencies_path(cluster_name)
}
\arguments{
\item{cluster_name}{Character. The name of the cluster. Capitolization and spacing
is ignored and does not matter (e.g., "My Cluster" is equivalent to "mycluster")}

\item{jars}{Character. A vector of jar files to include on the classpath. These should
be listed using the full file name, including the ".jar" extension at the end.}
}
\description{
Some cluster configuration packages contain java dependencies. This utility will
use the jar names from the configuration YAML and find the full file paths to the
stated dependencies.
}

