% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{is_cluster_known}
\alias{is_cluster_known}
\alias{list_available_clusters}
\title{Check to see if cluster has known configurations}
\usage{
is_cluster_known(cluster_name, install = "ask")

list_available_clusters()
}
\arguments{
\item{cluster_name}{Character. The name of the cluster. Capitolization and spacing
is ignored and does not matter (e.g., "My Cluster" is equivalent to "mycluster")}

\item{install}{Character. Should be one of \code{"ask"}, \code{"yes"}, or \code{"no"},
case insensitive. First letter abbreviations are suficient.}
}
\description{
This function will check to see if configurations for the named cluster exist. In
the even that configurations are available in a configured repository but not installed
locally then the user will be prompted to install the configurations.
}

