% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{is_cluster_known}
\alias{get_cluster_name}
\alias{get_cluster_package_name}
\alias{is_cluster_known}
\alias{list_available_clusters}
\title{Check to see if cluster has known configurations}
\usage{
is_cluster_known(cluster_name, install = "ask", stopifnot = FALSE)

list_available_clusters()

get_cluster_name(cluster_name)

get_cluster_package_name(cluster_name)
}
\arguments{
\item{cluster_name}{Character. The name of the cluster. Capitolization and spacing
is ignored and does not matter (e.g., "My Cluster" is equivalent to "mycluster")}

\item{install}{Character. Should be one of \code{"ask"}, \code{"yes"}, or \code{"no"},
case insensitive. First letter abbreviations are suficient.}

\item{stopifnot}{Logical. If \code{TRUE} then in the event a cluster is not known
an error will result. Otherwise that case will trigger a warning and a \code{FALSE}
return value.}
}
\description{
This function will check to see if configurations for the named cluster exist. In
the even that configurations are available in a configured repository but not installed
locally then the user will be prompted to install the configurations.
}

